move example specific utils back into example dirs
