+++ examples/vae/utils/vae_plots.py
