+++ examples/vae/vae_comparison.py
@@ -10,8 +10,8 @@ from torch.nn import functional
 from torchvision.utils import save_image
 
 import pyro
+from utils.mnist_cached import RESULTS_DIR, DATA_DIR
 from pyro.contrib.examples import util
-from pyro.contrib.examples.util import RESULTS_DIR
 from pyro.distributions import Bernoulli, Normal
 from pyro.infer import SVI
 from pyro.optim import Adam
@@ -211,10 +211,12 @@ class PyroVAEImpl(VAE):
 def setup(args):
     pyro.set_rng_seed(args.rng_seed)
     train_loader = util.get_data_loader(dataset_name='MNIST',
+                                        data_dir=DATA_DIR,
                                         batch_size=args.batch_size,
                                         is_training_set=True,
                                         shuffle=True)
     test_loader = util.get_data_loader(dataset_name='MNIST',
+                                       data_dir=DATA_DIR,
                                        batch_size=args.batch_size,
                                        is_training_set=False,
                                        shuffle=True)
